Moved db call to retrieve user conversations to ConversationWrapper
Relates #62
