+++ components/messaging/ConversationWrapper.tsx
@@ -1,17 +1,35 @@
 'use client';
 
-import { useContext } from 'react';
+import { useContext, useEffect } from 'react';
 import MeatballIcon from '../icons/MeatballIcon';
 import PlusIcon from '../icons/PlusIcon';
 import ConversationsList from './ConversationsList';
 import CurrentConversation from './CurrentConversation';
 import useConversation from '@/app/(dashboard)/conversations/useConversation';
 import useMediaQuery from '../hooks/useMediaQuery';
+import getUserConversationsandItemNames from '@/utils/messaging/getUserConversationsandItemNames';
 
-const ConversationWrapper: React.FC = () => {
+type ConversationWrapperType = {
+  userId: string;
+};
+
+const ConversationWrapper: React.FC<ConversationWrapperType> = ({ userId }) => {
   const isBreakpoint = useMediaQuery(1000);
-  const { showConversationsList, setShowConversationsList } =
-    useContext(useConversation);
+  const {
+    showConversationsList,
+    setShowConversationsList,
+    setAllConversations,
+  } = useContext(useConversation);
+
+  useEffect(() => {
+    const fetchConversations = async () => {
+      const fetchedConversations =
+        await getUserConversationsandItemNames(userId);
+
+      setAllConversations(fetchedConversations);
+    };
+    fetchConversations();
+  }, []);
   return (
     <>
       <div className='mt-4 flex justify-between px-3 '>
